requirements.txt
altgraph==0.17.5
cx-Freeze==5.1.1
dis3==0.1.3
future==1.0.0
pefile==2019.4.18
Pillow==6.2.2
pywin32-ctypes==0.2.0
simple-fstring==0.1
